Fix: arm64 type bugs — alarm ID, SCSI/RBF fields, inet stride, memcpy size

+++ Source/OS9exec_core/alarms.h
@@ -55,9 +55,9 @@
 void   init_alarms( void );
 void   A_Insert   ( alarm_typ* aa );
 void   A_Remove   ( alarm_typ* aa );
-os9err A_Make     ( ushort pid, ulong *aId, ushort aCode, ulong aTicks, Boolean cyclic );
+os9err A_Make     ( ushort pid, uint32_t *aId, ushort aCode, uint32_t aTicks, Boolean cyclic );
 void   A_Kill     ( ushort pid );
 
-os9err Alarm( ushort pid, ulong *aId, short aFunc, ushort aCode, ulong aTime, ulong aDate );
+os9err Alarm( ushort pid, uint32_t *aId, short aFunc, ushort aCode, uint32_t aTime, uint32_t aDate );
 
 /* eof */
